requirements.txt
# Argus - Semantic Search Service
# Extracted from Atlas

fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
sqlite-vec>=0.1.0
tiktoken>=0.5.0
requests>=2.31.0
pyyaml>=6.0
